post process LUT (wip)
* added a kernel based post process
* started working on a LUT post process
* post process can now be cycled back using shift + F9

+++ src/main/java/mayo/gui/widgets/Container.java
@@ -188,7 +188,7 @@ public class Container extends Widget implements Tickable, GUIListener {
 
         //change selected index
         i = backwards ? i - 1 : i + 1;
-        i = (int) Maths.modulo(i, list.size());
+        i = Maths.modulo(i, list.size());
 
         //return new selected widget
         return list.get(i);
